♻️  type 과 주석 담았습니다.
