+++ src/components/sales/calendar/cell/Cell.tsx
@@ -19,17 +19,27 @@ import { BIG_MODE, CALENDAR_PAGE, STATUS_PAGE } from '../calendarType/calendarTy
 import CellItem from '../cellItem/CellItem';
 import styles from './styles/cell.module.css';
 
-const Cell = ({ mode, page }: { mode: CalendarModeType; page?: CalendarPageType }) => {
-  const storeId = useAuthState(state => state.storeId);
-  const holidays = useHolidayState(state => state.holidays);
-  const calendarBindingData = useSalesDataState(state => state.calendarBindingData);
-
+const Cell = ({ mode, page }: CalendarCellType) => {
+  /**
+   * 캘린더에 사용되는 state입니다 건들지 마십쇼
+   */
   const currentDate = useCalendarState(state => state.currentDate);
-  const { clickShowDataOfDateHandler } = useDataHandler();
   // monthStart가 속한 주의 시작 주
   const startDay = currentDate.startOf('month').startOf('week');
   // monthStart가 속한 마지막 주
   const endDay = currentDate.endOf('month').endOf('week');
+
+  /**
+   * Sales Page에서 사용하는 state
+   */
+  const storeId = useAuthState(state => state.storeId);
+  const holidays = useHolidayState(state => state.holidays);
+  const calendarBindingData = useSalesDataState(state => state.calendarBindingData);
+  const { clickShowDataOfDateHandler } = useDataHandler();
+  const holiday = holidays[currentDate.format('YYYY')];
+  /**
+   * 다른 페이지에서 호출 되는 공간 입니다. 아래에서 hook으로 써주면 아리가또우
+   */
 
   useEffect(() => {
     if (mode === BIG_MODE) {
@@ -57,7 +67,7 @@ const Cell = ({ mode, page }: { mode: CalendarModeType; page?: CalendarPageType 
     }
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [currentDate]);
-  const holiday = holidays[currentDate.format('YYYY')];
+
   const row = [];
   let days = [];
   let day = startDay;
@@ -75,9 +85,14 @@ const Cell = ({ mode, page }: { mode: CalendarModeType; page?: CalendarPageType 
           day={day}
           salesData={salesData ? salesData[0] : undefined}
           // ... spreadOperator
+          // 아래 조건부 함수는 Sales page에서 사용하는 props 입니다.
           {...(page === CALENDAR_PAGE && { getMinMaxSalesType: getMinMaxSalesType })}
           {...(page === STATUS_PAGE && { clickShowDataOfDateHandler: clickShowDataOfDateHandler })}
           holiday={holidayDate}
+
+          /** 페이지가 주문내역 확인이면 아래와 같이 하면 됩니다.
+           *{...((page===ORDER && {onClick: clickHandler}))}
+           */
         />,
       );
       day = day.add(1, 'day');
