+++ src/components/sales/calendar/cellItem/CellItem.tsx
@@ -2,9 +2,7 @@ import { useModal } from '@/hooks/service/ui/useModal';
 import { convertNumberToWon } from '@/shared/helper';
 import useCalendarState from '@/shared/store/sales/salesCalendar';
 import useDayState from '@/shared/store/sales/salesDay';
-import { CalendarDataType, GetMinMaxSalesReturnType, HolidayType } from '@/types/sales';
 import { cva } from 'class-variance-authority';
-import { Dayjs } from 'dayjs';
 import {
   FORMAT_CELL_DATE_TYPE,
   getCalendarDateType,
@@ -16,18 +14,9 @@ import {
 } from '../../calendarUtility/cellItemType';
 import SalesModal from '../../modal/SalesModal';
 
-import { BIG_MODE, MINI_MODE, STATUS_PAGE } from '../calendarType/calendarType';
+import clsx from 'clsx';
+import { BIG_MODE, CALENDAR_PAGE, MINI_MODE, STATUS_PAGE } from '../calendarType/calendarType';
 import styles from './styles/cellItem.module.css';
-
-interface CellItemProps {
-  day: Dayjs;
-  salesData?: CalendarDataType;
-  getMinMaxSalesType?: (param: CalendarDataType) => GetMinMaxSalesReturnType;
-  clickShowDataOfDateHandler?: (day: Dayjs) => () => Promise<void>;
-  holiday: HolidayType[];
-  mode: CalendarModeType;
-  page?: CalendarPageType;
-}
 
 type Cell = (param: CellItemProps) => JSX.Element;
 
@@ -113,11 +102,15 @@ const CellItem: Cell = ({ day, salesData, getMinMaxSalesType, clickShowDataOfDat
       {/* sales/Status일 때 보여줄 날 css */}
       {mode === MINI_MODE && (
         <div
-          className={statusVariant({
-            calendarType: getStatusCalendarType(day, currentDate),
-            monthType: getStatusMonthType(currentDate, day),
-            dateType: getStatusDateType(day),
-            holidayType: holiday?.[0]?.name ? HOLIDAY : null,
+          className={clsx({
+            [statusVariant({
+              calendarType: getStatusCalendarType(day, currentDate),
+              monthType: getStatusMonthType(currentDate, day),
+              dateType: getStatusDateType(day),
+              holidayType: holiday?.[0]?.name ? HOLIDAY : null,
+            })]: page === STATUS_PAGE,
+
+            // [styles.페이지가 주문내역이면 사용하는 style] : PAGE === ORDER_PAGE
           })}
           {...(((page === STATUS_PAGE && day.isSame(today, 'D')) || day.isBefore(today, 'D')) && {
             onClick: clickShowDataOfDateHandler?.(day),
@@ -141,11 +134,14 @@ const CellItem: Cell = ({ day, salesData, getMinMaxSalesType, clickShowDataOfDat
       {/* sales/Calendar일 때 보여줄 날 css */}
       {mode === BIG_MODE && (
         <div
-          className={calendarVariant({
-            monthType: getCalendarMonthType(day, currentDate),
-            dateType: getCalendarDateType(day),
-            salesType: salesData ? SALES_HAVE : SALES_NONE,
-          })}
+          className={clsx(
+            page === CALENDAR_PAGE &&
+              calendarVariant({
+                monthType: getCalendarMonthType(day, currentDate),
+                dateType: getCalendarDateType(day),
+                salesType: salesData ? SALES_HAVE : SALES_NONE,
+              }),
+          )}
           {...(day.format(FORMAT_CELL_DATE_TYPE) === salesData?.date && {
             onClick: () => {
               MagicModal.fire(<SalesModal specificData={salesData!} />);
